+++ src/components/modules/paste.ts
@@ -2,15 +2,14 @@ import Module from '../__module';
 import $ from '../dom';
 import * as _ from '../utils';
 import {
-  BlockTool,
-  BlockToolConstructable,
-  PasteConfig,
+  BlockAPI,
   PasteEvent,
   PasteEventDetail
 } from '../../../types';
 import Block from '../block';
 import { SavedData } from '../../../types/data-formats';
 import { clean, sanitizeBlocks } from '../utils/sanitizer';
+import BlockTool from '../tools/block';
 
 /**
  * Tag substitute object.
@@ -19,9 +18,8 @@ interface TagSubstitute {
   /**
    * Name of related Tool
    *
-   * @type {string}
    */
-  tool: string;
+  tool: BlockTool;
 }
 
 /**
@@ -30,24 +28,18 @@ interface TagSubstitute {
 interface PatternSubstitute {
   /**
    * Pattern`s key
-   *
-   * @type {string}
    */
   key: string;
 
   /**
    * Pattern regexp
-   *
-   * @type {RegExp}
    */
   pattern: RegExp;
 
   /**
    * Name of related Tool
-   *
-   * @type {string}
    */
-  tool: string;
+  tool: BlockTool;
 }
 
 /**
@@ -229,7 +221,7 @@ export default class Paste extends Module {
    * @param {boolean} isHTML - if passed string is HTML, this parameter should be true
    */
   public async processText(data: string, isHTML = false): Promise<void> {
-    const { Caret, BlockManager, Tools } = this.Editor;
+    const { Caret, BlockManager } = this.Editor;
     const dataToInsert = isHTML ? this.processHTML(data) : this.processPlain(data);
 
     if (!dataToInsert.length) {
@@ -246,7 +238,7 @@ export default class Paste extends Module {
       return;
     }
 
-    const isCurrentBlockDefault = BlockManager.currentBlock && Tools.isDefault(BlockManager.currentBlock.tool);
+    const isCurrentBlockDefault = BlockManager.currentBlock && BlockManager.currentBlock.tool.isDefault;
     const needToReplaceCurrentBlock = isCurrentBlockDefault && BlockManager.currentBlock.isEmpty;
 
     dataToInsert.map(
@@ -278,23 +270,22 @@ export default class Paste extends Module {
   private processTools(): void {
     const tools = this.Editor.Tools.blockTools;
 
-    Object.entries(tools).forEach(this.processTool);
+    Array
+      .from(tools.values())
+      .forEach(this.processTool);
   }
 
   /**
    * Process paste config for each tool
+   *
+   * @param tool - BlockTool object
    */
-  private processTool = ([name, tool]: [string, BlockToolConstructable]): void => {
+  private processTool = (tool: BlockTool): void => {
     try {
-      const toolInstance = new this.Editor.Tools.blockTools[name]({
-        api: this.Editor.API.getMethodsForTool(name),
-        config: {},
-        data: {},
-        readOnly: false,
-      }) as BlockTool;
+      const toolInstance = tool.create({}, {} as BlockAPI, false);
 
       if (tool.pasteConfig === false) {
-        this.exceptionList.push(name);
+        this.exceptionList.push(tool.name);
 
         return;
       }
@@ -303,14 +294,12 @@ export default class Paste extends Module {
         return;
       }
 
-      const toolPasteConfig = tool.pasteConfig || {};
-
-      this.getTagsConfig(name, toolPasteConfig);
-      this.getFilesConfig(name, toolPasteConfig);
-      this.getPatternsConfig(name, toolPasteConfig);
+      this.getTagsConfig(tool);
+      this.getFilesConfig(tool);
+      this.getPatternsConfig(tool);
     } catch (e) {
       _.log(
-        `Paste handling for «${name}» Tool hasn't been set up because of the error`,
+        `Paste handling for «${tool.name}» Tool hasn't been set up because of the error`,
         'warn',
         e
       );
@@ -320,17 +309,16 @@ export default class Paste extends Module {
   /**
    * Get tags to substitute by Tool
    *
-   * @param {string} name - Tool name
-   * @param {PasteConfig} toolPasteConfig - Tool onPaste configuration
+   * @param tool - BlockTool object
    */
-  private getTagsConfig(name: string, toolPasteConfig: PasteConfig): void {
-    const tags = toolPasteConfig.tags || [];
+  private getTagsConfig(tool: BlockTool): void {
+    const tags = tool.pasteConfig.tags || [];
 
     tags.forEach((tag) => {
       if (Object.prototype.hasOwnProperty.call(this.toolsTags, tag)) {
         _.log(
-          `Paste handler for «${name}» Tool on «${tag}» tag is skipped ` +
-          `because it is already used by «${this.toolsTags[tag].tool}» Tool.`,
+          `Paste handler for «${tool.name}» Tool on «${tag}» tag is skipped ` +
+          `because it is already used by «${this.toolsTags[tag].tool.name}» Tool.`,
           'warn'
         );
 
@@ -338,21 +326,20 @@ export default class Paste extends Module {
       }
 
       this.toolsTags[tag.toUpperCase()] = {
-        tool: name,
+        tool,
       };
     });
 
-    this.tagsByTool[name] = tags.map((t) => t.toUpperCase());
+    this.tagsByTool[tool.name] = tags.map((t) => t.toUpperCase());
   }
 
   /**
    * Get files` types and extensions to substitute by Tool
    *
-   * @param {string} name - Tool name
-   * @param {PasteConfig} toolPasteConfig - Tool onPaste configuration
+   * @param tool - BlockTool object
    */
-  private getFilesConfig(name: string, toolPasteConfig: PasteConfig): void {
-    const { files = {} } = toolPasteConfig;
+  private getFilesConfig(tool: BlockTool): void {
+    const { files = {} } = tool.pasteConfig;
     let { extensions, mimeTypes } = files;
 
     if (!extensions && !mimeTypes) {
@@ -360,19 +347,19 @@ export default class Paste extends Module {
     }
 
     if (extensions && !Array.isArray(extensions)) {
-      _.log(`«extensions» property of the onDrop config for «${name}» Tool should be an array`);
+      _.log(`«extensions» property of the onDrop config for «${tool.name}» Tool should be an array`);
       extensions = [];
     }
 
     if (mimeTypes && !Array.isArray(mimeTypes)) {
-      _.log(`«mimeTypes» property of the onDrop config for «${name}» Tool should be an array`);
+      _.log(`«mimeTypes» property of the onDrop config for «${tool.name}» Tool should be an array`);
       mimeTypes = [];
     }
 
     if (mimeTypes) {
       mimeTypes = mimeTypes.filter((type) => {
         if (!_.isValidMimeType(type)) {
-          _.log(`MIME type value «${type}» for the «${name}» Tool is not a valid MIME type`, 'warn');
+          _.log(`MIME type value «${type}» for the «${tool.name}» Tool is not a valid MIME type`, 'warn');
 
           return false;
         }
@@ -381,7 +368,7 @@ export default class Paste extends Module {
       });
     }
 
-    this.toolsFiles[name] = {
+    this.toolsFiles[tool.name] = {
       extensions: extensions || [],
       mimeTypes: mimeTypes || [],
     };
@@ -390,19 +377,18 @@ export default class Paste extends Module {
   /**
    * Get RegExp patterns to substitute by Tool
    *
-   * @param {string} name - Tool name
-   * @param {PasteConfig} toolPasteConfig - Tool onPaste configuration
+   * @param tool - BlockTool object
    */
-  private getPatternsConfig(name: string, toolPasteConfig: PasteConfig): void {
-    if (!toolPasteConfig.patterns || _.isEmpty(toolPasteConfig.patterns)) {
+  private getPatternsConfig(tool: BlockTool): void {
+    if (!tool.pasteConfig.patterns || _.isEmpty(tool.pasteConfig.patterns)) {
       return;
     }
 
-    Object.entries(toolPasteConfig.patterns).forEach(([key, pattern]: [string, RegExp]) => {
+    Object.entries(tool.pasteConfig.patterns).forEach(([key, pattern]: [string, RegExp]) => {
       /** Still need to validate pattern as it provided by user */
       if (!(pattern instanceof RegExp)) {
         _.log(
-          `Pattern ${pattern} for «${name}» Tool is skipped because it should be a Regexp instance.`,
+          `Pattern ${pattern} for «${tool.name}» Tool is skipped because it should be a Regexp instance.`,
           'warn'
         );
       }
@@ -410,7 +396,7 @@ export default class Paste extends Module {
       this.toolsPatterns.push({
         key,
         pattern,
-        tool: name,
+        tool,
       });
     });
   }
@@ -461,7 +447,7 @@ export default class Paste extends Module {
    * @param {FileList} items - pasted or dropped items
    */
   private async processFiles(items: FileList): Promise<void> {
-    const { BlockManager, Tools } = this.Editor;
+    const { BlockManager } = this.Editor;
 
     let dataToInsert: {type: string; event: PasteEvent}[];
 
@@ -472,7 +458,7 @@ export default class Paste extends Module {
     );
     dataToInsert = dataToInsert.filter((data) => !!data);
 
-    const isCurrentBlockDefault = Tools.isDefault(BlockManager.currentBlock.tool);
+    const isCurrentBlockDefault = BlockManager.currentBlock.tool.isDefault;
     const needToReplaceCurrentBlock = isCurrentBlockDefault && BlockManager.currentBlock.isEmpty;
 
     dataToInsert.forEach(
@@ -529,7 +515,6 @@ export default class Paste extends Module {
    */
   private processHTML(innerHTML: string): PasteData[] {
     const { Tools } = this.Editor;
-    const initialTool = this.config.defaultBlock;
     const wrapper = $.make('DIV');
 
     wrapper.innerHTML = innerHTML;
@@ -538,7 +523,7 @@ export default class Paste extends Module {
 
     return nodes
       .map((node) => {
-        let content, tool = initialTool, isBlock = false;
+        let content, tool = Tools.defaultTool, isBlock = false;
 
         switch (node.nodeType) {
           /** If node is a document fragment, use temp wrapper to get innerHTML */
@@ -558,7 +543,7 @@ export default class Paste extends Module {
             break;
         }
 
-        const { tags } = Tools.blockTools[tool].pasteConfig as PasteConfig;
+        const { tags } = tool.pasteConfig;
 
         const toolTags = tags.reduce((result, tag) => {
           result[tag.toLowerCase()] = {};
@@ -576,7 +561,7 @@ export default class Paste extends Module {
         return {
           content,
           isBlock,
-          tool,
+          tool: tool.name,
           event,
         };
       })
@@ -626,7 +611,7 @@ export default class Paste extends Module {
    * @param {PasteData} dataToInsert - data of Block to inseret
    */
   private async processSingleBlock(dataToInsert: PasteData): Promise<void> {
-    const { Caret, BlockManager, Tools } = this.Editor;
+    const { Caret, BlockManager } = this.Editor;
     const { currentBlock } = BlockManager;
 
     /**
@@ -637,7 +622,7 @@ export default class Paste extends Module {
       dataToInsert.tool !== currentBlock.name ||
       !$.containsOnlyInlineElements(dataToInsert.content.innerHTML)
     ) {
-      this.insertBlock(dataToInsert, currentBlock && Tools.isDefault(currentBlock.tool) && currentBlock.isEmpty);
+      this.insertBlock(dataToInsert, currentBlock?.tool.isDefault && currentBlock.isEmpty);
 
       return;
     }
@@ -657,14 +642,14 @@ export default class Paste extends Module {
     const { BlockManager, Caret, Tools } = this.Editor;
     const { content } = dataToInsert;
 
-    const currentBlockIsDefault = BlockManager.currentBlock && Tools.isDefault(BlockManager.currentBlock.tool);
+    const currentBlockIsDefault = BlockManager.currentBlock && BlockManager.currentBlock.tool.isDefault;
 
     if (currentBlockIsDefault && content.textContent.length < Paste.PATTERN_PROCESSING_MAX_LENGTH) {
       const blockData = await this.processPattern(content.textContent);
 
       if (blockData) {
         const needToReplaceCurrentBlock = BlockManager.currentBlock &&
-          Tools.isDefault(BlockManager.currentBlock.tool) &&
+          BlockManager.currentBlock.tool.isDefault &&
           BlockManager.currentBlock.isEmpty;
 
         const insertedBlock = BlockManager.paste(blockData.tool, blockData.event, needToReplaceCurrentBlock);
@@ -677,7 +662,7 @@ export default class Paste extends Module {
 
     /** If there is no pattern substitute - insert string as it is */
     if (BlockManager.currentBlock && BlockManager.currentBlock.currentInput) {
-      const currentToolSanitizeConfig = Tools.getInlineToolsSanitizeConfigForBlock(BlockManager.currentBlock.name);
+      const currentToolSanitizeConfig = Tools.getInlineToolsSanitizeConfigForBlock(BlockManager.currentBlock.tool);
 
       document.execCommand(
         'insertHTML',
@@ -718,7 +703,7 @@ export default class Paste extends Module {
 
     return {
       event,
-      tool: pattern.tool,
+      tool: pattern.tool.name,
     };
   }
 
@@ -762,7 +747,7 @@ export default class Paste extends Module {
       let needToReplaceCurrentBlock = false;
 
       if (i === 0) {
-        const isCurrentBlockDefault = BlockManager.currentBlock && Tools.isDefault(BlockManager.currentBlock.tool);
+        const isCurrentBlockDefault = BlockManager.currentBlock && BlockManager.currentBlock.tool.isDefault;
 
         needToReplaceCurrentBlock = isCurrentBlockDefault && BlockManager.currentBlock.isEmpty;
       }
@@ -791,8 +776,8 @@ export default class Paste extends Module {
 
     const element = node as HTMLElement;
 
-    const { tool = '' } = this.toolsTags[element.tagName] || {};
-    const toolTags = this.tagsByTool[tool] || [];
+    const { tool } = this.toolsTags[element.tagName] || {};
+    const toolTags = this.tagsByTool[tool?.name] || [];
 
     const isSubstitutable = tags.includes(element.tagName);
     const isBlockElement = $.blockElements.includes(element.tagName.toLowerCase());