+++ src/components/modules/toolbar/blockSettings.ts
@@ -144,10 +144,10 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
     /**
      * Add default settings that presents for all Blocks
      */
-    this.addDefaultSettings();
+    this.addTunes();
 
     /** Tell to subscribers that block settings is opened */
-    this.Editor.Events.emit(this.events.opened);
+    this.eventsDispatcher.emit(this.events.opened);
 
     this.flipper.activate(this.blockTunesButtons);
   }
@@ -183,7 +183,7 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
     this.nodes.defaultSettings.innerHTML = '';
 
     /** Tell to subscribers that block settings is closed */
-    this.Editor.Events.emit(this.events.closed);
+    this.eventsDispatcher.emit(this.events.closed);
 
     /** Clear cached buttons */
     this.buttons = [];
@@ -229,16 +229,21 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
    * Add Tool's settings
    */
   private addToolSettings(): void {
-    if (_.isFunction(this.Editor.BlockManager.currentBlock.tool.renderSettings)) {
-      $.append(this.nodes.toolSettings, this.Editor.BlockManager.currentBlock.tool.renderSettings());
+    const settingsElement = this.Editor.BlockManager.currentBlock.renderSettings();
+
+    if (settingsElement) {
+      $.append(this.nodes.toolSettings, settingsElement);
     }
   }
 
   /**
-   * Add default settings
+   * Add tunes: provided by user and default ones
    */
-  private addDefaultSettings(): void {
-    $.append(this.nodes.defaultSettings, this.Editor.BlockManager.currentBlock.renderTunes());
+  private addTunes(): void {
+    const [toolTunes, defaultTunes] = this.Editor.BlockManager.currentBlock.renderTunes();
+
+    $.append(this.nodes.toolSettings, toolTunes);
+    $.append(this.nodes.defaultSettings, defaultTunes);
   }
 
   /**