+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 2.20.0
+
+- `New` — [Block Tunes API](block-tunes.md) added
+
+### 2.19.3
+
+- `Fix` — Ignore error raised by Shortcut module
+
 ### 2.19.2
 
 - `New` - `toolbar.toggleBlockSettings()` API method added [#1442](https://github.com/codex-team/editor.js/issues/1421).
@@ -15,6 +23,7 @@
 - `Fix` - Fix unexpected behavior on an empty link pasting [#1348](https://github.com/codex-team/editor.js/issues/1348).
 - `Fix` - Fix SanitizerConfig type definition [#1513](https://github.com/codex-team/editor.js/issues/1513)
 - `Refactoring` - The Listeners module now is a util.
+- `Refactoring` - The Events module now is a util.
 - `Fix` - Editor Config now immutable [#1552](https://github.com/codex-team/editor.js/issues/1552).
 - `Refactoring` - The Sanitizer module is util now.
 - `Refactoring` - Shortcuts module is util now.